+++ webapp/tests/unit/log/log-doctypes-caching.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit tests for log docTypes caching and enhanced logging functionality
  * @description     Tests for docTypes caching, enhanced log format, and admin logs search functionality
  * @file            webapp/tests/unit/log/log-doctypes-caching.test.js
- * @version         1.6.32
- * @release         2026-03-21
+ * @version         1.6.33
+ * @release         2026-03-22
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/